refactor(registry): fix Zeitwerk naming convention
- Split manifest.rb into registry_manifest.rb (RegistryManifest) and
  pack_definition.rb (PackDefinition) so filenames match Zeitwerk expectations
- Remove require_relative calls from registry.rb — Zeitwerk autoloads all
  registry/ files correctly with proper naming
- Add parse_yaml guards: rescue Psych::SyntaxError → ParseError; reject
  non-Hash YAML as ParseError
- Add specs for invalid YAML syntax and non-mapping YAML frontmatter
- Rename spec files to match new source file names
- Add trailing newline to docs/port-registry-resolution.md
- Update plan doc file list to reflect renamed files
